chore: clear the small lint categories in test/
explicit-types (uint -> uint256), a missing SPDX header and pragma in
_GatewayPreparationForTests, and per-site exceptions with reasons for forge-std's
lowercase `vm` handle, a string literal that embeds double quotes, and the
da-contracts shim interface whose name deliberately mirrors the contract it
stands in for.

`reentrancy` is off for tests: solhint's heuristic fires on ordinary
state-then-call sequencing in harness code, where it carries no signal.

+++ l1-contracts/test/foundry/l1/unit/concrete/Governance/ServerNotifier.t.sol
@@ -80,7 +80,7 @@ contract ServerNotifierTest is Test {
     }
 
     function test_setUpgradeTimestampValidProtocolVersionSucceeds() public {
-        uint deadline = block.timestamp + 7 days;
+        uint256 deadline = block.timestamp + 7 days;
 
         chainTypeManager.setUpgradeCutHash(protocolVersion, keccak256("upgradeCutHash"));
         chainTypeManager.setProtocolVersionDeadline(protocolVersion, deadline);
@@ -95,7 +95,7 @@ contract ServerNotifierTest is Test {
 
     function test_setUpgradeTimestampCutDataForProtocolVersionNotAvailableReverts() public {
         chainTypeManager.setUpgradeCutHash(protocolVersion, bytes32(0));
-        uint deadline = block.timestamp + 7 days;
+        uint256 deadline = block.timestamp + 7 days;
 
         chainTypeManager.setProtocolVersionDeadline(protocolVersion, deadline);
 
@@ -105,7 +105,7 @@ contract ServerNotifierTest is Test {
     }
 
     function test_setUpgradeTimestampInvalidCallerReverts() public {
-        uint deadline = block.timestamp + 7 days;
+        uint256 deadline = block.timestamp + 7 days;
 
         chainTypeManager.setProtocolVersionDeadline(protocolVersion, deadline);
